+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Versioning practices: [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
  - Check for 'rsync' before starting to collect statistics.
  - Add support for workloads that provide custom lables. Use the '--pipe' option
    for that. The 'stc-wl-cpu-wake-walk' is an example of such a workload.
+ - Add hover text to turbostat scatter plots.
 ### Removed
 ### Changed